+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+* Field order is preserved in serialized documents
+
 1.0.0
 -----